Laveau smack not working. Started fixes
The implementation for the sandal and locomotion look as though they working as intended so it may be an issue with the blend between smack and idle/run or possibly an issue with the animation. I need Brandon's help to get it to work. Still need to add matchbox, lighter fluid and syringe animations.